Updated DeckController: Updated Paths and Responses of multiple requests -Need to work on Update to actually update the deck contents, a placeholder string output has been made to show the expected shuffled deck
Updated DeckDao
Updated methods so they all are able to work. There have consistency issues due to lack of events being used for file creation, update, and deletion. Creation and deletion will work with multiple attempts, but the event handling is missing, update will not work. A placeholder string output in the terminal will show the shuffled contents that should be written into the file.

Updated Deck
Updated the writing methods so it would write into the correct directory.